Updates to EOA checks
Skip verified EOA check if tx.origin is "to" to minimize friction for users when purchasing NFTs on approved marketplaces.

Use tx.origin instead of a message signature to flag an account as an EOA in the EOARegistry. For a transaction to be submitted to the network it must be signed by an EOA, adding a second signature is redundant.